Move //java/testing/javac to //java/util/javac
Also extract :package_util
