+++ java/jflex/migration/TestCaseVars.java
@@ -1,6 +1,8 @@
 package jflex.migration;
 
-public class TestCaseVars {
+import jflex.velocity.TemplateVars;
+
+public class TestCaseVars extends TemplateVars {
 
   public String javaPackage;
   public String testCaseName;
