+++ java/jflex/testing/testsuite/BUILD
@@ -8,8 +8,8 @@ java_library(
         "//java/jflex/testing/testsuite/annotations",
     ],
     deps = [
-        "//java/jflex/testing/javac",
         "//java/jflex/testing/testsuite/annotations",
+        "//java/jflex/util/javac",
         "//jflex",
         "//jflex/src/main/java/jflex/generator",
         "//third_party/com/google/guava",
