+++ java/jflex/testing/testsuite/JFlexTestRunner.java
@@ -32,10 +32,10 @@ import static org.junit.Assert.fail;
 import com.google.common.collect.ImmutableList;
 import java.io.File;
 import jflex.generator.LexGenerator;
-import jflex.testing.javac.CompilerException;
-import jflex.testing.javac.JavacUtil;
 import jflex.testing.testsuite.annotations.NoExceptionThrown;
 import jflex.testing.testsuite.annotations.TestSpec;
+import jflex.util.javac.CompilerException;
+import jflex.util.javac.JavacUtil;
 import org.junit.runner.Description;
 import org.junit.runner.notification.Failure;
 import org.junit.runner.notification.RunNotifier;
