+++ java/jflex/ucd_generator/emitter/unicode_version/UnicodeVersionEmitter.java
@@ -8,10 +8,10 @@ import java.io.Writer;
 import java.util.Iterator;
 import java.util.SortedSet;
 import java.util.stream.Collectors;
-import jflex.testing.javac.PackageUtil;
 import jflex.ucd_generator.emitter.common.UcdEmitter;
 import jflex.ucd_generator.scanner.UnicodeData;
 import jflex.ucd_generator.ucd.UcdVersion;
+import jflex.util.javac.JavaPackageUtil;
 import jflex.velocity.Velocity;
 import org.apache.velocity.runtime.parser.ParseException;
 
@@ -19,7 +19,7 @@ import org.apache.velocity.runtime.parser.ParseException;
 public class UnicodeVersionEmitter extends UcdEmitter {
 
   private static final String UNICODE_VERSION_TEMPLATE =
-      PackageUtil.getPathForClass(UnicodeVersionEmitter.class) + "/Unicode_x_y.java.vm";
+      JavaPackageUtil.getPathForClass(UnicodeVersionEmitter.class) + "/Unicode_x_y.java.vm";
   private static final String CP_ZERO = escapedUTF16Char(0);
 
   private final UcdVersion ucdVersion;
