--- javatests/jflex/testing/javac/BUILD
@@ -1,8 +1,0 @@
-java_test(
-    name = "PackageUtilTest",
-    srcs = ["PackageUtilTest.java"],
-    deps = [
-        "//java/jflex/testing/javac",
-        "//third_party/com/google/truth",
-    ],
-)
