+++ javatests/jflex/util/javac/BUILD
@@ -1,0 +1,8 @@
+java_test(
+    name = "JavaPackageUtilTest",
+    srcs = ["JavaPackageUtilTest.java"],
+    deps = [
+        "//java/jflex/util/javac:package_util",
+        "//third_party/com/google/truth",
+    ],
+)
